chore: remove GetJSONSerializableMap
